+++ 0x05-pointers_arrays_strings/main.h
@@ -1,8 +1,10 @@
 #ifndef MAIN_H
 #define MAIN_H
 
+#include <stdlib.h>
+
 void reset_to_98(int *c);
 void swap_int(int *a, int *b);
-int strlen(const char *str);
+size_t strlen(const char *str);
 
 #endif
